Add support for MPTCP filtering
"mptcp" can now be used to filter TCP packets that contain an MPTCP option.
mptcp is used as a protocol qualifier (as are tcp, udp, etc.). The code is
generic enough to filter on any kind of TCP options if required.

One issue with filtering on MPTCP packets is that the BPF code is large.
Indeed, we need to iterate over each TCP option and as BPF only allow to
jump forward we need to duplicate the BPF code for each potential TCP option.

Signed-off-by: Gregory Detal <gregory.detal@uclouvain.be>